WIP Restructuring; currently symmetric solve isn't working
