+++ .directory
@@ -1,0 +1,3 @@
+[Dolphin]
+Timestamp=2018,12,19,22,26,0
+Version=4
